+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Change log
 ### 4.13.0 (8/4/2017 - )
 * web: Response.file(File) changed to Response.file(Path) to be consistent with entire api design, (to return path via WebDirectory)  
+* json: update jackson to 2.9.0, simplified configuration
 
 ### 4.12.7 (8/1/2017 - 8/3/2017)
 * es: update to 5.5.1