+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Change log
-### 4.13.0 (8/4/2017 - )
+### 4.13.0 (8/4/2017 - 8/11/2017)
 * web: Response.file(File) changed to Response.file(Path) to be consistent with entire api design, (to return path via WebDirectory)  
 * json: update jackson to 2.9.0, simplified configuration
 * scheduler: add secondly trigger, to support precisely scheduling to align with clock time